+++ src/index.js
@@ -69,12 +69,16 @@ export default class DI {
         return !(typeof this.services[name] === 'undefined');
     }
 
+    addParameter(name, value) {
+        if (this.parameters.hasOwnProperty(name)) {
+            throw new Error("Parameter '" + name + "' already exists");
+        }
+        this.parameters[name] = value;
+    }
+
     addParameters(parameters) {
         for (let parameterName in parameters) {
-            if (this.parameters.hasOwnProperty(parameterName)) {
-                throw new Error("Parameter '" + parameterName + "' already exists");
-            }
-            this.parameters[parameterName] = parameters[parameterName];
+            this.addParameter(parameterName, parameters[parameterName]);
         }
     }